requirements.txt
langchain==0.1.4
openai==1.10.0
pypdf==4.0.1
faiss-cpu==1.7.4
python-dotenv==1.0.0
tiktoken==0.5.2
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6 